Update WeightedGraph.java
Fixed addWeightedEdge to allow for edges to go to from one Node to another Node which has an edge to the first Node.
Added that each Node would be provided the set of all Nodes in the graph when created.
Added method to get distance from one Ndoe to another.