Code Review resubmission
Added comments into the code in areas where reading was difficult to follow.
Fixed question_5.Node to not have each Node contain a variable pointing to a set of all Nodes in the graph it was in.
Fixed question_5.Main to work with the new change on question_5.Node.
Changed question_6.GridGraph so that undirectedEdgeExists is used in removeUndirectedEdge and made removeUndirectedEdge throw an Exception if it was provided a GridNode not in the graph.
Changed question_6.Main.astar to use backtracking to figure out what path was taken upon finding the ending rather than do the weird thing I had before where I would put all GridNodes into retList, then at the end remove the ones I didn't need.
Changed question_7.Main.main to generate larger graphs to compare dijkstras and astar. Also added some print statements to better see when certain areas of the code were being done.